+++ test_runner/src/test/kotlin/ftl/test/util/TestArtifact.kt
@@ -114,7 +114,6 @@ object TestArtifact {
 
         val unzip = "unzip \"$zipPath\" -d \"${fixtures.path}\""
         Bash.execute(unzip)
-        File(zipPath).copyTo(File("${fixtures.path}/EarlGreyExample.zip"))
     }
 }